+++ Core.hpp
@@ -5,7 +5,7 @@
 // Login   <erwan.simon@epitech.eu>
 // 
 // Started on  Sat Jan 21 13:15:51 2017 erwan
-// Last update Sat Jan 21 21:16:07 2017 erwan
+// Last update Sat Jan 21 22:40:19 2017 Pierre-Emmanuel Merlier
 //
 
 #ifndef CORE_HPP_
@@ -25,7 +25,7 @@ public:
 public:
   std::string	getCPUModel() const;
   int		getCoreNb() const;
-  float		*getCorePercent() const;
+  long long	*getCorePercent() const;
   float		*getRam() const;
   float		*getSwap() const;
   int		getNbTasks() const;
@@ -33,7 +33,7 @@ public:
 
   void		setCPUModel(std::string CPUModel);
   void		setCoreNb(int core);
-  void		setCorePercent(float *corePercent);
+  void		setCorePercent(long long *corePercent);
   void		setRam(float *ram);
   void		setSwap(float *swap);
   void		setNbTasks(int tasksNb);
@@ -41,7 +41,7 @@ public:
 private:
   std::string	_CPUModel;
   int		_coreNb;
-  float		*_corePercent; // tableau contenant dans l'ordre des cores leur pourcentage d'utilisation
+  long long	*_corePercent; // tableau contenant dans l'ordre des cores leur pourcentage d'utilisation
   float		*_ram;
   float		*_swap;
   int		_nbTasks; // nb tache active
